requirements.txt
paho-mqtt
influxdb